pyproject.toml
[project]
name = "v2ray-aggregator"
version = "1.0.0"
description = "V2Ray 免费节点聚合系统 - 多源爬取、测速筛选、多格式输出"
authors = [
    {name = "V2Ray Team"}
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "aiohttp>=3.9.1",
    "requests>=2.31.0",
    "PyYAML>=6.0.1",
    "pydantic>=2.5.0",
    "aiofiles>=23.2.1",
    "async-timeout>=4.0.3",
    "python-dotenv>=1.0.0",
    "colorlog>=6.8.0",
    "tqdm>=4.66.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "utils"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

